fix: do not commit erroneous state in `akp_cluster`
In the case an AKP API error happens, this should not be committed
into TF state. This is a bit more involved than for `akp_instance`
since it is e.g. possible for a cluster to be created in AKP, but
the manifests for whatever reason may not be created (when
applicable).