fix: filter cm fields not already present in resource (#319)
This would fix a lot of the inconsistent state issues which we see,
where if we don't specify the exact keys that the API returns, the
provider fails horribly. This also simplifies the config from a user
perspective, since they don't have to specify every single key that the
API returns. This means that from a TF perspective, the config becomes
minimal in most cases.

This relies on the fact that when we submit requests to the ApplyConfig
endpoint, that only the cm fields specified are set; any other field
that needs to be present in the data model relies on the API to set the
correct defaults.

Fixes #143